run the smoketests in parallel
* capture all the subprocess pids
* for each smoketest, wait and add up returns, as before

The one thing that is slightly grotty about this change is that all the
smoketest output will be intermingled.  That might be addressed, if we
don't like it, by having each smoke test collect output as it goes then
blurt it all out in one call at the end.  Locally things look clean and
understandable, but this is a typical issue when making these kinds of
changes.