use sponge to linearize reports
sponge is in moreutils `brew install moreutils` and here is used to soak
up all the stdout from a test, only printing that output when the test
script exits.

smoketest then pipes each test run to sponge, using process substitution
so we can still grab the pid of the test itself as before.  Then before
we print the final result, we wait again for any straggler sponges.