wait for kubectl and es-console
I think with the change to making tests parallel we exposed this test a
little, in that it assumes es-console is ready by the time this test
starts.  Now that all the tests start at the same time, we should wrap
our work steps in busy wait and gives things as long as they need to
start up.