Quicker smoke (#68)
* lighter test app

This uses es-test instead of es-demo.  es-test is a tiny distroless go
app that does nothing but produce "ohai" metrics for prometheus.
es-demo is an akka app and has been seen to take over a minute to start
on Travis.  Hopefully this is faster!

* give tls ingress unique name

This let's us create and test both tls and non-tls ingress at the same
time without worrying about them stepping on each other.

* move start test apps to top

Slightly faster, hopefully.

* run the smoketests in parallel

* capture all the subprocess pids
* for each smoketest, wait and add up returns, as before

The one thing that is slightly grotty about this change is that all the
smoketest output will be intermingled.  That might be addressed, if we
don't like it, by having each smoke test collect output as it goes then
blurt it all out in one call at the end.  Locally things look clean and
understandable, but this is a typical issue when making these kinds of
changes.

* use sponge to linearize reports

sponge is in moreutils `brew install moreutils` and here is used to soak
up all the stdout from a test, only printing that output when the test
script exits.

smoketest then pipes each test run to sponge, using process substitution
so we can still grab the pid of the test itself as before.  Then before
we print the final result, we wait again for any straggler sponges.

* wait for kubectl and es-console

I think with the change to making tests parallel we exposed this test a
little, in that it assumes es-console is ready by the time this test
starts.  Now that all the tests start at the same time, we should wrap
our work steps in busy wait and gives things as long as they need to
start up.

* wait for kube-state-metrics and node-exporter

instead of waiting for Prometheus.  If we're scraping those two then the
data set in this test is in hand.

* debug travis

* ubuntu ps format fix

on unbutu, ps -opid=,pgid=command= means "set the column header for pid
to be ",pgid=command=" and just prints pid.  So this never worked on
Travis.