tweaks on credential file, install-local
Cinnamon documents this file with a dot, not an underbar. This changes
to be consistent with that file name.

Also changes `install-local` to install from the folder contents instead
of a tarball, as what you want to install are your local file changes
and not some old tarball by mistake.  This also eliminates the constant
rewriting of a previous release.