tweaks on credential file, install-local (#97)
* tweaks on credential file, install-dev

Cinnamon documents this file with a dot, not an underbar. This changes
to be consistent with that file name.

Also adds `install-dev` to install from the folder contents instead
of a tarball, as what you want to install are your local file changes
and not some old tarball by mistake.  This also eliminates the constant
rewriting of a previous release.

* install-dev target suitable for local dev loop

* busy_wait test_es_console_responding

* don't flunk CI if --wait fails