resilient nginx (#114)
* resilient nginx

1. service name lookups are not critical errors, but rather just cause
nginx to serve 502's until the service name is available, at which point
nginx will switch over to it automatically.  This fixes the issue of
nginx crashing if prometheus has not yet started, as well as fixes the
issue of nginx failing if prometheus is moved.

2. base paths are dynamic.  However you arrive at nginx, it preserves
the base path for the Console SPA, Grafana, Prometheus, and
es-monitor-api.  So one user can be using a port-forward while another
is using an ingress while another is using kube proxy, and it'll all
just work.

3. busts browser caching for the Console SPA entry point (index.html) as
well as our Grafana dashboard plugin (exporter-async.js). This should
stop the problem of browser's caching older versions and not pickup up
updates.

4. routes Grafana session cookies properly, so session store doesn't get
stuffed with new sessions for every request

5. all services are directly reachable, so /service/prometheus in the
browser will open prometheus and it'll just work

as well as some miscellaneous installer cleanup, having picked up a bad merge from master...