relabel container and pod metrics with workload
Following the naming conventions used for Pod Template Hashes and the
SimpleNameGenerator, as well as the reduced SafeString character set, we
pattern match for workload names out of pods with these patterns:

1. workload-hash-simple
2. workload-simple
3. workload-number
4. workload

While not perfect, it seems to work quite well in the main. Is is
certainly possible to incorrectly match "hash" or "simple" with plain
names that have those patterns, but given the size constraints and the
fact that vowels are not allowed reduces the change of error.

With this we also swapped in correct metrics for the Grafana Kubernetes
section, since we can match container stats with ContextTags.

fixes lightbend/es-backend#310