fix prometheus javascript for double proxy (#121)
Going through kubectl proxy to our nginx to prometheus results in double
rewriting of the links. The outer proxy might have a URI like:

http://127.0.0.1:8001/api/v1/namespaces/lightbend/services/console-server/proxy/service/prometheus/graph

which is then turns into a request to nginx for this path:

/service/prometheus/graph

which then turns into a request on prometheus for:

/graph

This is all working fine, as links at the nginx layer are then rewritten
to have /service/prometheus base paths and the kubectl proxy then
rewrites them again to have the full /api/v1.../proxy/ base path.

However!  Prometheus has one funky piece of javascript that tries to
rewrite the base path as well :rolls-eyes: but as you can see in the
above scenario our nginx may not be in a position to know the final
path. So, this changes the javascript path builder to just look at the
current browser location, whatever it may be, and take off /graph to get
to the /static things that prometheus UI needs to load.

double rainbow all the way!