invert "free" to "used" for per-process viewpoint
Free per process is confusing. We want to say what portion a process is
using rather than what portion a process is not using, as the former is
simple to understand at 3am with pagerduty alerting, while the later is
hard to understand even if you stare at it quietly with a little
worksheet on the side to do the math.

There's probably something clever we can do to put all the context
together in a single view to provide a window into resource exhaustion
at the container level, but until then I'll just go with this.

Added Active Memory Used to better zoom in on typical JVM memory leak
use case.