Make prometheus data mount writeable
PVCs can get mounted as 755, so only root users can write to them.
We can set the fsGroup on the pod security context to handle this.

Unfortunately container level security contexts don't affect volume
mounts. So this fsGroup applies to all containers in the pod.

As all the other containers run as the root user, this should be okay.

This may not work in Openshift - we need to verify that.