Run smoketests_prometheus on central park
This involved quite a bit of rework of the existing tests and Makefile:

* Moved the install logic/running logic into scripts, so it's easier to
share code between minikube/openshift and understand. This also
simplifies the Makefile a bit.

* Use a custom annotation domain for tests so we don't conflict with
other things installed in Central Park.

* Renamed expose-es-console to expose-console-server so it is
consistently named with the console-server Service.

* Made NAMESPACE parameterized for all the tests, which were hardcoded
to 'lightbend' or 'smoketest-prometheus' in many cases.

* The test resources/console/tiller are all installed in the same
namespace, so I can more easily control the setup/teardown, as well as
simplifying the openshift deployment (where the service account only has
access to a single namespace).

* Simplified the ingress tests a bit - they were incorrectly using the
ndoeport of the service, when they can just use port 80 on the
console-server service.

* Fixed 'nodeport' function to work on openshift with routes.

* Added throttling to travis debug log output, to try and prevent the
travis build from dying/OOMKilled as keeps happening. I'm not sure this
will help, but worth a try. Since I added it, I haven't had a flaky
build in my PR.

* Move helm install into setup-tools-for-ubuntu.sh since we need it for
both minikube and openshift.

* Fix the helm version so it doesn't randomly break, as we are relying
on a pre-setup Tiller on Central Park.

* Fix frontend e2e to not depend on hardcoded values - so it will work
with the above changes.

* Add some diagnostics to frontend e2e so its easier to debug them if
the environment is broken.

* Always upload artifacts even on success.  So we can easily see what a
good run looks like, when comparing to a failed run.