Fix helm upgrade on an old release
That was using the rolling update strategy. This fixes the helm
error that occurs when doing an upgrade.

Error: UPGRADE FAILED: Deployment.apps "prometheus-alertmanager" is
invalid: spec.strategy.rollingUpdate: Forbidden: may not be specified
when strategy `type` is 'Recreate' && Deployment.apps "grafana-server"
is invalid: spec.strategy.rollingUpdate: Forbidden: may not be specified
when strategy `type` is 'Recreate' && Deployment.apps
"prometheus-server" is invalid: spec.strategy.rollingUpdate: Forbidden:
may not be specified when strategy `type` is 'Recreate'