Use cluster DNS directly in nginx config (#258)
* Use cluster DNS directly in nginx config

Instead of using go-dnsmasq, just parse /etc/resolv.conf directly and
insert the needed values into nginx.conf.

This avoids the need for an intermediate container, and relies on the
stable interface of /etc/resolv.conf. Note there is currently no way to
query k8s for this information other than checking resolv.conf.

For https://github.com/lightbend/es-backend/issues/541

* Fix alertmanager substitution

We can't arbitrarily add KUBE_DOMAIN, because if a user wants to
integrate with an external alertmanager they will provide the full DNS
resolution here. So instead, add KUBE_DOMAIN in values.yaml.

* Rename to KUBE_CONSOLE_DOMAIN

So it is clear this is the console domain (in the namespace of the
console install), rather than the generic cluster domain (which doesn't
include namespace).

* Fix init container to work when podUID is set

* Remove extra blank line

* Skip tiller log output - it's too noisy

* Just use sed so we don't need to install a package

On Openshift and when running as non-root, the init container can't
install packages at runtime.