Reduce rate periods in static-rules (#292)
* Reduce rate periods in static-rules

* Having very large periods can extrapolate single errors for a long
time. This leads to a poor experience when it seems to take a long time
for a monitor to go ready. An example is the akka_http_server_5xx
monitor, which will go orange for 5m+ with only a few 5xx at the start
up of the cluster.
* We should rely on our monitor models for handling spikey data via its
window/confidence mechanism, rather than smoothing things out in static
rules.

It's also not great we have these rules - they are hidden, if a user is
just looking at the UI. We should think of how we can surface this
better. I think they should just be dynamically generated recording
rules, ideally.

* Use irates so it is safer with longer scrape times