Add redirect to trailing slash
This change checks for the presence of a single level path without a
trailing slash, and then redirects it.

This should work in all cases, because we already require that all
application routes have specific location blocks to handle them.

Also some other small changes:
- Added anchors to all the regex matches, as recommended by nginx best
practices. This just makes sure we don't match things we shouldn't.
- Dropped 'index index.html' in the final catch-all location block. I
couldn't figure out the point of this, and it didn't seem to have any
effect removing it.