Queue travis builds (#335)
Queue a single travis build at a time

It uses a golang script `travisqueue.go` which queries the travis API for any builds
in progress. If there is a build in progress that started before the current one,
it will poll until that condition is no longer met (all prior builds have finished executing).

The `travisqueue.go` is build based, so can be placed directly on the jobs as the first
script step. This lets us protect against people retrying builds as well.

It should be placed an any job with shared state. Currently this is the openshift
gotests and the deploy job.

Also other build improvements:
* Don't perform unnecessary install steps in jobs that don't need it
* travis_retry gotest setup to reduce flakiness