handle both jmx_exporter and cinnamon jvm metrics
fixes https://github.com/lightbend/console-backend/issues/646

If jmx_exporter version of jvm metrics is available for a workload,
we'll use those. If not then we'll try the Cinnamon version.

Some of the Cinnamon metrics have a GC or Pool value embedded in the
name, so here we parse those out into a tag to get the same ergonomics
as jmx_exporter. This keeps the jmx_exporter or cinnamon_metric
expression lined up, compresses multiple promql expressions into one,
and handles any gc or pool type found without having to know them ahead
of time.

Note that for GC Rate and GC Time we are using Prometheus 2.7 subquery
features so promtool and in-cluster prometheus must be >= 2.7 for these
to work.